fixed stuff according to mr
